refactor: 파일구조 변경
fix: 계좌 기본 제공 금액 변경
